Move local `getFullPageExamples()` lib function into review app
It’s not shared and is only used locally
