+++ packages/govuk-frontend-review/package.json
@@ -24,6 +24,7 @@
     "cookie-parser": "^1.4.6",
     "express": "^4.18.2",
     "express-validator": "^7.0.1",
+    "front-matter": "^4.0.2",
     "govuk-frontend": "*",
     "govuk-frontend-config": "*",
     "govuk-frontend-lib": "*",
