+++ shared/lib/files.unit.test.js
@@ -1,4 +1,4 @@
-const { getComponentData, getFullPageExamples } = require('./files.js')
+const { getComponentData } = require('./files.js')
 
 describe('getComponentData', () => {
   it('rejects if unable to load component data', async () => {
@@ -43,44 +43,5 @@ describe('getComponentData', () => {
         data: expect.any(Object)
       })
     ))
-  })
-})
-
-describe('getFullPageExamples', () => {
-  it('contains name of each example', async () => {
-    const examples = await getFullPageExamples()
-
-    examples.forEach((example) => expect(example).toEqual(
-      expect.objectContaining({
-        name: expect.any(String),
-        path: expect.any(String)
-      })
-    ))
-  })
-
-  it('contains scenario front matter, for some examples', async () => {
-    const examples = await getFullPageExamples()
-
-    // At least 1x example with { name, path, scenario }
-    expect(examples).toEqual(expect.arrayContaining([
-      expect.objectContaining({
-        name: expect.any(String),
-        path: expect.any(String),
-        scenario: expect.any(String)
-      })
-    ]))
-  })
-
-  it('contains notes front matter, for some examples', async () => {
-    const examples = await getFullPageExamples()
-
-    // At least 1x example with { name, path, notes }
-    expect(examples).toEqual(expect.arrayContaining([
-      expect.objectContaining({
-        name: expect.any(String),
-        path: expect.any(String),
-        notes: expect.any(String)
-      })
-    ]))
   })
 })
